Extract SPDX license JSON parsing to its own class
This allows testing in isolation. And as a result of testing, fix a bug where we were not allowing the 'https' variant of license URLs.

This also changes the behavior around duplicated license URLs to prefer the shortest identifier for now. At the very least this is probably the one you want, and it makes the behavior deterministic rather than relying on the JSON order.