Support Kotlin multiplatform + application plugin
The application plugin applies the java plugin which was causing a collision. Instead of sometimes configuring our plugin when other plugins were available and sometimes configuring it in afterEvaluate, just always do the work in afterEvaluate. This ensures we can synchronously query the availability of plugins and make decisions with normal control flow.